refactor: Service 레이어 삭제
